Extract RadioButton
